Database collections are now returned as streams whenever possible

+++ httpserver/src/main/kotlin/pt/isel/ps/g06/httpserver/controller/IngredientController.kt
@@ -16,6 +16,7 @@ import pt.isel.ps.g06.httpserver.service.IngredientService
 import pt.isel.ps.g06.httpserver.service.UserService
 import javax.validation.constraints.Max
 import javax.validation.constraints.Min
+import kotlin.streams.toList
 
 @Validated
 @RestController
@@ -36,8 +37,8 @@ class IngredientController(
      */
     @GetMapping
     fun getIngredients(
-            @RequestParam @Min(0) skip: Int?,
-            @RequestParam @Min(0) @Max(MAX_COUNT) count: Int?
+            @RequestParam @Min(0) skip: Long?,
+            @RequestParam @Min(0) @Max(MAX_COUNT) count: Long?
     ): ResponseEntity<IngredientsContainerOutput> {
 
         val ingredients = ingredientService.getIngredients(skip, count)
@@ -50,7 +51,6 @@ class IngredientController(
     /**
      * Create a hardcoded ingredient as moderator
      * @param   user    ModAuthorizationArgumentResolver's parameter
-     * @param   ingredientInput    The ingredient to be inserted
      * @return  [ResponseEntity]<[IngredientsContainerOutput]>
      * @throws InvalidInputException On invalid cuisines passed
      */
